// src/lib/feeder-gateway/feeder-gateway-service-type.ts
enum FeederGatewayServiceType {
  IS_READY = 'is_ready',
  IS_ALIVE = 'is_alive',
  GET_BATCH_ENCLOSING_IDS = 'get_batch_enclosing_ids',
  GET_BATCH_IDS = 'get_batch_ids',
  GET_BATCH_INFO = 'get_batch_info',
  GET_LAST_BATCH_ID = 'get_last_batch_id',
  GET_PREV_BATCH_ID = 'get_prev_batch_id'
}

export {FeederGatewayServiceType};